Rename game to demo
