+++ src/screens/playing.rs
@@ -3,7 +3,7 @@
 use bevy::{input::common_conditions::input_just_pressed, prelude::*};
 
 use super::Screen;
-use crate::game::spawn::level::SpawnLevel;
+use crate::demo::spawn::level::SpawnLevel;
 use crate::{assets::SoundtrackHandles, audio::soundtrack::SoundtrackCommands as _};
 
 pub(super) fn plugin(app: &mut App) {
